remove casting output of dot

+++ transformer_engine/jax/dot.py
@@ -27,7 +27,7 @@ def type_safe_dot_general(
     if fp8_meta_pkg is None:
         assert x.dtype == kernel.dtype, f"lhs dtype = {x.dtype}, rhs dtype = {kernel.dtype}"
         output = jax.lax.dot_general(x, kernel, (contracting_dims, ((), ())))
-        return output.astype(x.dtype)
+        return output
 
     amax_list = fp8_meta_pkg.amax_list
     scale_list = fp8_meta_pkg.scale_list
